+++ src/components/shopping_cart/Table.jsx
