+++ src/components/shopping_cart/TicketRow.jsx
